feat: reboot & refactor mmio
